Add files via upload
Fixed bug in stacks where only the first and last colours were stored.
Added the ability to print the stacks this will be implemented in a couple of places.
Added the ability to remove from the bottom of the stack after a predefined number. It is currently set to 2 for testing purposes but will be changed to 5 in the final project.